+++ scripts/upgrade.py
@@ -43,6 +43,7 @@ SUPPORTED_SOURCES = {
     "1.0.0",
     "1.0.1",
     "1.0.2",
+    "1.0.3",
     installer.INSTALLER_VERSION,
 }